refactor: create VertexGenAi provider and refactor GenAi initialization
