+++ app/src/genAi/types.ts
@@ -1,0 +1,20 @@
+import { AiPrompt, AiPromptResponse } from "../types";
+
+export interface GenAiConfig {
+  apiEndpoint: string;
+  projectId: string;
+  locationId: string;
+  modelId: string;
+  maxOutputTokens: number;
+  systemInstruction?: string;
+  membersInstruction?: string;
+  safetySettings?: {
+    category: string;
+    threshold: string;
+  }[];
+}
+
+export interface GenAi {
+  init(): Promise<void>;
+  generate(prompt: AiPrompt): Promise<AiPromptResponse>;
+}
